handling new email equally between service and app
