+++ NotificationExtension/NotificationService.swift
@@ -7,32 +7,40 @@
 //
 
 import UserNotifications
-import SwiftSoup
+import RealmSwift
 
 class NotificationService: UNNotificationServiceExtension {
 
     var contentHandler: ((UNNotificationContent) -> Void)?
     var bestAttemptContent: UNMutableNotificationContent?
-    let database = Database()
-    let apiRequest = APIRequest()
+    
+    override init() {
+        super.init()
+        
+        let fileURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: "group.criptext.team")!.appendingPathComponent("default.realm")
+        let config = Realm.Configuration(
+            fileURL: fileURL,
+            schemaVersion: Env.databaseVersion)
+        Realm.Configuration.defaultConfiguration = config
+    }
 
     override func didReceive(_ request: UNNotificationRequest, withContentHandler contentHandler: @escaping (UNNotificationContent) -> Void) {
         self.contentHandler = contentHandler
         bestAttemptContent = (request.content.mutableCopy() as? UNMutableNotificationContent)
         guard let bestAttemptContent = bestAttemptContent,
-            let account = database.getFirstAccount() else {
+            let account = SharedDB.getFirstAccount() else {
                 contentHandler(request.content)
                 return
         }
-        apiRequest.getEvents(token: account.jwt) { (events) in
+        SharedAPI.getEvents(token: account.jwt) { (responseData) in
             let userInfo = request.content.userInfo
-            guard let myEvents = events,
+            guard case let .SuccessArray(events) = responseData,
                 let keyString = userInfo["metadataKey"] as? String,
                 let key = Int(keyString) else {
                 contentHandler(request.content)
                 return
             }
-            self.handleEvents(myEvents, for: key) { responseEmail in
+            self.handleEvents(events, for: key) { responseEmail in
                 guard let email = responseEmail else {
                     contentHandler(request.content)
                     return
@@ -61,69 +69,9 @@ class NotificationService: UNNotificationServiceExtension {
             completion(nil)
             return
         }
-        handleNewEmailCommand(params: event, completion: completion)
-    }
-    
-    func handleNewEmailCommand(params: [String: Any], completion: @escaping (_ email: Email?) -> Void){
-        guard let myAccount = database.getFirstAccount() else {
-            completion(nil)
-            return
-        }
-        let event = NewEmail.init(params: params)
-        
-        let email = Email()
-        email.threadId = event.threadId
-        email.subject = event.subject
-        email.key = event.metadataKey
-        email.messageId = event.messageId
-        email.date = event.date
-        email.unread = true
-        
-        apiRequest.getEmailBody(metadataKey: email.key, token: myAccount.jwt) { (bodyResponse) in
-            guard let myAccount = self.database.getFirstAccount(),
-                let body = bodyResponse,
-                let username = ServiceContactUtils.getUsernameFromEmailFormat(event.from),
-                let content = self.handleBodyByMessageType(event.messageType, body: body, account: myAccount, recipientId: username, senderDeviceId: event.senderDeviceId) else {
-                completion(nil)
-                return
-            }
-            
-            let contentPreview = self.getContentPreview(content: content)
-            email.content = contentPreview.1
-            email.preview = contentPreview.0
-            
-            guard self.database.store(email) else {
-                completion(nil)
-                return
-            }
-            
-            ServiceContactUtils.parseEmailContacts([event.from], database: self.database, email: email, type: .from)
-            ServiceContactUtils.parseEmailContacts(event.to, database: self.database, email: email, type: .to)
-            ServiceContactUtils.parseEmailContacts(event.cc, database: self.database, email: email, type: .cc)
-            ServiceContactUtils.parseEmailContacts(event.bcc, database: self.database, email: email, type: .bcc)
-            
-            completion(email)
-        }
-    }
-    
-    func handleBodyByMessageType(_ messageType: MessageType, body: String, account: Account, recipientId: String, senderDeviceId: Int32?) -> String? {
-        guard messageType != .none,
-            let deviceId = senderDeviceId else {
-                return body
-        }
-        return SignalHandler.decryptMessage(body, messageType: messageType, account: account, recipientId: recipientId, deviceId: deviceId)
-    }
-    
-    func getContentPreview(content: String) -> (String, String) {
-        do {
-            let allowList = try SwiftSoup.Whitelist.relaxed().addTags("style", "title", "header").addAttributes(":all", "class", "style", "src")
-            let doc: Document = try SwiftSoup.parse(content)
-            let preview = try String(doc.text().prefix(100))
-            let cleanContent = try SwiftSoup.clean(content, allowList)!
-            return (preview, cleanContent)
-        } catch {
-            let preview = String(content.prefix(100))
-            return (preview, content)
+        let newEmailHandler = NewEmailHandler()
+        newEmailHandler.command(params: event) { (result) in
+            completion(result.email)
         }
     }
