+++ iOS-Email-Client/Libs/Env.swift
@@ -30,14 +30,16 @@ struct Env {
     }
     
     static var socketURL: String {
-        return "wss://stagesocket.criptext.com:3002"
+        return "wss://socket.criptext.com:3002"
     }
     
     static var domain: String {
-        return "@jigl.com"
+        return "@criptext.com"
     }
     
     static var apiURL: String {
-        return "https://stage.mail.criptext.com"
+        return "https://api.criptext.com"
     }
+    
+    static let databaseVersion: UInt64 = 7
 }
